+++ local_test/db/docker/db/init/10_property_authentication.sql
