+++ local_test/db/docker/db/init/11_property_authentication_deleted.sql
