--- property-author-update/function.json
@@ -1,11 +1,0 @@
-{
-	"scriptFile": "../dist/property-ahtuor-update/index.js",
-	"bindings": [
-		{
-			"name": "myTimer",
-			"type": "timerTrigger",
-			"direction": "in",
-			"schedule": "0 12-57/15 * * * *"
-		}
-	]
-}
