--- property-factory-change-author/function.json
@@ -1,11 +1,0 @@
-{
-	"scriptFile": "../dist/property-factory-change-author/index.js",
-	"bindings": [
-		{
-			"name": "myTimer",
-			"type": "timerTrigger",
-			"direction": "in",
-			"schedule": "0 11-56/15 * * * *"
-		}
-	]
-}
