+++ test/main/common/db/dao.ts
@@ -1,3 +1,4 @@
+import { Connection } from 'typeorm'
 import { DbConnection, Transaction } from '../../../../common/db/common'
 import {
 	EventTableAccessor,
@@ -6,10 +7,18 @@ import {
 	getEventRecordThenGreaterBlockNumber,
 	getProcessedBlockNumber,
 	setProcessedBlockNumber,
+	PropertyBalanceAccessor,
 } from '../../../../common/db/dao'
 import { LockupLockedup } from '../../../../entities/lockup-lockedup'
+import { PropertyBalance } from '../../../../entities/property-balance'
 import { getDbConnection } from '../../../lib/db'
-import { saveLockupLockupedTestdata, clearData } from '../../../lib/test-data'
+import {
+	generateTestAddress,
+	saveLockupLockupedTestdata,
+	clearData,
+	getCount,
+	EventDataGenerator,
+} from '../../../lib/test-data'
 
 describe('EventTableAccessor', () => {
 	let con: DbConnection
@@ -154,5 +163,153 @@ describe('getProcessedBlockNumber,setProcessedBlockNumber', () => {
 		await transaction.finish()
 		const blockNumber = await getProcessedBlockNumber(con.connection, 'test')
 		expect(blockNumber).toBe(999)
+	})
+})
+
+describe('PropertyBalanceAccessor', () => {
+	let con: DbConnection
+	const [
+		property1,
+		property2,
+		property3,
+		account1,
+		account2,
+		account3,
+		account4,
+		treasury,
+	] = generateTestAddress()
+	async function saveTestData1(con: Connection) {
+		const transaction = new Transaction(con)
+		await transaction.start()
+		const record = new PropertyBalance()
+		record.property_address = property1
+		record.account_address = account1
+		record.balance = '100'
+		record.is_author = true
+		record.block_number = 10000
+		await transaction.save(record)
+		record.account_address = account2
+		record.balance = '1000'
+		record.is_author = false
+		record.block_number = 10000
+		await transaction.save(record)
+		record.property_address = property2
+		record.account_address = account3
+		record.balance = '1000'
+		record.is_author = true
+		record.block_number = 100000
+		await transaction.save(record)
+		record.account_address = account4
+		record.balance = '1000000'
+		record.is_author = false
+		record.block_number = 1000000
+		await transaction.save(record)
+		await transaction.commit()
+		await transaction.finish()
+	}
+
+	beforeAll(async (done) => {
+		con = await getDbConnection()
+		done()
+	})
+	afterAll(async (done) => {
+		await con.quit()
+		done()
+	})
+	beforeEach(async () => {
+		await clearData(con.connection, PropertyBalance)
+	})
+	describe('deleteRecord', () => {
+		it('delete records that match the specified address.', async () => {
+			await saveTestData1(con.connection)
+			const transaction = new Transaction(con.connection)
+			await transaction.start()
+			const accessor = new PropertyBalanceAccessor(transaction)
+			await accessor.deleteRecord(property1)
+			await transaction.commit()
+			await transaction.finish()
+			const repository = con.connection.getRepository(PropertyBalance)
+			const records = await repository.find({
+				property_address: property2,
+			})
+			expect(records.length).toBe(2)
+		})
+		it('if there is no matching record, the record will not be deleted.', async () => {
+			await saveTestData1(con.connection)
+			const transaction = new Transaction(con.connection)
+			await transaction.start()
+			const accessor = new PropertyBalanceAccessor(transaction)
+			await accessor.deleteRecord(property3)
+			await transaction.commit()
+			await transaction.finish()
+			const recordCount = await getCount(con.connection, PropertyBalance)
+			expect(recordCount).toBe(4)
+		})
+	})
+	describe('insertRecord', () => {
+		it('event information is registered as a record..', async () => {
+			await saveTestData1(con.connection)
+			const dataGenerator = new EventDataGenerator()
+			dataGenerator.addMintTransfer(account1, 10000, 100)
+			dataGenerator.addMintTransfer(treasury, 5000, 100)
+			dataGenerator.addTransfer(account1, account4, 100, 120)
+			dataGenerator.addTransfer(account1, account3, 50, 122)
+			const transaction = new Transaction(con.connection)
+			await transaction.start()
+			const accessor = new PropertyBalanceAccessor(transaction)
+			await accessor.insertRecord(dataGenerator.data, property1, account1)
+			await transaction.commit()
+			await transaction.finish()
+			const repository = con.connection.getRepository(PropertyBalance)
+			const property2Records = await repository.find({
+				property_address: property2,
+			})
+			expect(property2Records.length).toBe(2)
+			const property1Records = await repository.find({
+				property_address: property1,
+			})
+			expect(property1Records.length).toBe(4)
+			property1Records.forEach((property1Record) => {
+				expect(property1Record.property_address).toBe(property1)
+				switch (property1Record.account_address) {
+					case treasury:
+						expect(property1Record.balance).toBe('5000')
+						expect(property1Record.is_author).toBe(false)
+						expect(property1Record.block_number).toBe(100)
+						break
+					case account4:
+						expect(property1Record.balance).toBe('100')
+						expect(property1Record.is_author).toBe(false)
+						expect(property1Record.block_number).toBe(120)
+						break
+					case account3:
+						expect(property1Record.balance).toBe('50')
+						expect(property1Record.is_author).toBe(false)
+						expect(property1Record.block_number).toBe(122)
+						break
+					case account1:
+						expect(property1Record.balance).toBe('9850')
+						expect(property1Record.is_author).toBe(true)
+						expect(property1Record.block_number).toBe(122)
+						break
+					default:
+						throw new Error('illegal account address')
+				}
+			})
+		})
+		it('if the event information is zero, an error occurs.', async () => {
+			const dataGenerator = new EventDataGenerator()
+			const transaction = new Transaction(con.connection)
+			await transaction.start()
+			const accessor = new PropertyBalanceAccessor(transaction)
+			const result = await accessor
+				.insertRecord(dataGenerator.data, property1, account1)
+				.catch((err: Error) => err)
+			expect((result as Error).message).toBe(
+				`property balance record is 0: ${property1}`
+			)
+			await transaction.commit()
+			await transaction.finish()
+		})
 	})
 })
